requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
anthropic==0.18.1
schedule==1.2.0
python-dotenv==1.0.0
feedparser==6.0.10
lxml==5.1.0
playwright==1.41.0
gunicorn==21.2.0
sendgrid==6.11.0
